update snakeyaml dependency to hard requirement
Jeseme had a "soft requirement" on 1.26. Dependabot build logs show an automatic usage of 1.31 and any version prior to 1.31 is affected by a security issue. So, the requirement was changed to a hard requirement on any version being 1.31 or greater (https://maven.apache.org/pom.html#Dependency_Version_Requirement_Specification)